requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
python-dotenv
sqlalchemy
aiosqlite
greenlet
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4
langchain-core
langchain-text-splitters
langgraph
numpy
pypdf
python-docx
reportlab
# ==========================================
# Mac M4 (Apple Silicon) AI/NLP Environment
# ==========================================

# 1. 核心科学计算与机器学习 (Core ML & Data)
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 2. 深度学习框架 (Deep Learning - Auto-enables MPS on M4)
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# 3. 前沿 NLP 与大模型底层 (Modern NLP & LLM Backend)
transformers>=4.38.0
datasets>=2.18.0
tokenizers>=0.15.0
accelerate>=0.27.0    # 优化 M4 内存加载与设备分配的核心库
sentence-transformers # 用于生成文本向量 (Embeddings)

# 4. Agent 与复杂工作流生态 (LangChain & LangGraph Ecosystem)
langchain>=0.1.10
langchain-core>=0.1.30
langchain-community>=0.0.25
langgraph>=0.0.26
langchain-openai      # 常用 LLM 接口
langsmith             # 调试和监控 LangChain 工作流

# 5. Web 服务与 API 构建 (API & Web Frameworks)
fastapi>=0.110.0      # 高性能 API 框架，完美搭配 LangChain
uvicorn[standard]     # 运行 FastAPI 的 ASGI 服务器

# 6. 开发与调试工具 (Dev Tools)
jupyterlab
tqdm
python-dotenv         # 用于管理 API Keys (如 .env 文件)
faiss-cpu
pymilvus
pinecone
